Fix reversed operators support
The sanitizer didn't work when chaining conditions and the operators
were reversed. It was because the sanitizer didn't change the status
when OR & AND keywords where present in the query to
`Status::ComparisonScopeStarted`.

Now these two new options are evaluated when sanitizing and the correct
status is set up so it knows when to put a placeholder on sensitive
values.