Remove trailing comments
It's common for the database adapter or its instrumentation to add
a comment to the query with contextual information, such as the
database instance that it was performed again or the trace ID.

This removes comments located at the very end of the trace, which
is where these comments are usually found. Other comments are
preserved, as they often contain user-provided context about the
query itself.