Respect dotted quoted identifiers in parenthesis
When a double-quoted value is found, it is unclear whether to treat
it as a string (MySQL) or as an identifier (standard SQL), so the
placeholder logic must treat it as a string. However, if it is
preceded or followed by a dot, then we know it must be an identifier,
and we should not replace it with a placeholder.