Fix SolidCache JSON sanitisation
The issue lies on the fact that a binary MySQL string (`x"..."`)
appears before a regular MySQL string (`"..."`). This trips the
sanitiser, which recognises the leading `x` as a literal value type
indicator. After the literal value type indicator state is over,
the sanitiser forgets that it's inside a VALUES statement.

Fix this by removing the literal value type indicator state and
keeping the state as-is when encountering the literal value type
indicator token.

Co-authored-by: Luismi Ramirez <luismi@appsignal.com>