Remove most function arguments
Removes function arguments in most contexts. Leaves some function
arguments in scenarios where they are unambiguously column names.

Treat `WHERE` as a comparison operator. Treat `LIMIT` and `FROM`
as switching to the default (non-sanitizable) context. Treat `IN`
as a keyword, as it is likely followed by a list literal which we'll
want to sanitise.

Only treat unknown keywords as keywords -- in practice limiting the
sanitisable keyword scope to function arguments.

Remove the `ComparisonScopeStarted` state, instead expanding what
is treated as a comparison operator.

This changes the existing sanitisation of some queries -- see the
changes to the tests.